sms dispatching machinery is started when the server starts
this is needed since the hooking strategy for ancel needs the driver to be loaded when the contacts and groups are being saved. if we were to wait for an sms notification to start then we would probably be losing many contacts-to-be since the driver wasn't loaded when the administraction action occured.